file digest with hash + sig
