+++ package.json
@@ -241,7 +241,8 @@
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^5.0.9",
-    "worker-loader": "^3.0.6"
+    "worker-loader": "^3.0.6",
+    "yargs": "^17.1.1"
   },
   "dependencies": {
     "@grpc/grpc-js": "^1.3.6",
